Agent: allow tests to be shared
